Multiple changes and bug fix in dodge prob and critical hit
Bug fix in dodge prob and critical hit
Since we forget to cast either numerator or denominator to double, it always returns zero.
So we cast the numerator to double first.

Fix in showing max/min attack of monster

Add showing the drop weapon status if when the weapon inventory is full to give info to player to decide which weapon he should drop.

Showing whether the monster is flying or not when consuming fruit, providing info to player to decide which weapon he should use.